lmkd: export basic statistics to sys.lmk.count
Tracking how much lmkd kills were on the system for a period of time is
an important piece of metric to keep track of.

These were hard to get in its native form. Export it to sys.lmk.count
in module_param_array() fashion.

Counts are logged every 10 kills to micro-optimize against
property_set()'s overhead.

Note that sepolicy rule must be changed in order to adb to retrieve
these statistics without root.

Change-Id: Id16e4cbbdb5c1da0c9dd719bebc0570f274561d6
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>